Mandrake patches from   http://cvs.mandrakesoft.com/cgi-bin/cvsweb.cgi/SPECS/xine-lib/ via Goetz Waschk who reports:   The amd64 patch (xine-lib-1-rc0a-amd64.patch) sets some conservative   CFLAGS for amd64,   the lib64 patch (xine-lib-1-rc0a-lib64.patch) replaces hardcoded   /lib to support the lib64 library dir on amd64,   the directfb patch (xine-lib-1-rc2-no-directfb.patch) adds a   configure option to disable directfb,   the linuxfb patch (xine-lib-1-rc3a-no-linuxfb.patch) does the same   for linux framebuffer and   the 64bit fixes patch (xine-lib-1-rc3-64bit-fixes.patch) doesn't   apply at the moment against the CVS -- demux_ogg.c was not applied.   it includes some 64 bit pointer and other fixes for 64bit architectures.   from Gwenole Beauchesne I haven't tested other than apply and compile.
CVS patchset: 6174
CVS date: 2004/02/19 02:50:24

+++ src/post/goom/filters.c
@@ -18,6 +18,7 @@
 #include <stdlib.h>
 #include <math.h>
 #include <stdio.h>
+#include <stdint.h>
 
 #ifdef HAVE_MMX
 #define USE_ASM
@@ -559,15 +560,15 @@ zoomFilterFastRGB (Uint * pix1,
 
 			freebrutS =
 				(unsigned int *) calloc (resx * resy * 2 + 128, sizeof(unsigned int));
-			brutS = (guint32 *) ((1 + ((unsigned int) (freebrutS)) / 128) * 128);
+			brutS = (guint32 *) ((1 + ((uintptr_t) (freebrutS)) / 128) * 128);
 
 			freebrutD =
 				(unsigned int *) calloc (resx * resy * 2 + 128, sizeof(unsigned int));
-			brutD = (guint32 *) ((1 + ((unsigned int) (freebrutD)) / 128) * 128);
+			brutD = (guint32 *) ((1 + ((uintptr_t) (freebrutD)) / 128) * 128);
 
 			freebrutT =
 				(unsigned int *) calloc (resx * resy * 2 + 128, sizeof(unsigned int));
-			brutT = (guint32 *) ((1 + ((unsigned int) (freebrutT)) / 128) * 128);
+			brutT = (guint32 *) ((1 + ((uintptr_t) (freebrutT)) / 128) * 128);
 
 			/** modif here by jeko : plus de multiplications **/
 			{
